optr: Fix GetSafeLockName function, add tests (#6876)
Fix the `GetSafeLockName()` function that makes a lock name safe as a
k8s resource name, and rejecting where it is invalid and cant be made
safe. In particular, it was mistranslating digits into hyphens,
translating periods into hyphens and allowing non-ASCII characters.

Add tests to cover the full range of the function. There have been a few
bugs in the function so a comprehensive set of tests here is desirable.

Source-commit: gravitational/teleport.e@416d44f1a34d791d6ff3567249674e0cebea2f15
Source-pull-requests: gravitational/teleport.e#6876