tooling/oprt: Wait for k8s lease end before exiting (#6824)
Synchronize canceling the k8s lease with waiting for the k8s lease
goroutine to finish, so that we do not exit the program too early before
the leader elector has a chance to update the k8s resource indicating it
has been released.

Currently the code path on completion cancels the context that the
leaderElector is running on before returning, but there is little code
after that before the process will exit. While that little code runs,
the leader elector needs to communicate with the k8s api server to
modify a k8s lease resource to indicate it has been released. If this
does not complete before main completes, the program will exit without
releasing the lock.

By having the lease cancel function return a channel that we wait on
that signals the lease has been released, it will be safe to continue
process termination without risk of leaving the lease locked.

Source-commit: gravitational/teleport.e@f0e029f71df82ccdd125aa374169aec38fd4e711
Source-pull-requests: gravitational/teleport.e#6824