[AWSIC] Adds provisioning event hooks to the SCIM Provisioner (#6742)
In order for the enclosing Identity Center plugin to be able to suppress downstream
group deletion in certain circumstances (see issue #6660), the IC integartion needs
to know when (de)provisioning events are happening and be offered the opportunity to
suppress the (de)provisioning if necessary.

This change adds hooks that the IC intentration will use. The idea is to keep IC-specific
logic _out_ of the generic SCIM provisioning system, while still delegating provisioning
decisions to the Identity Center integration.

This patch also adds/refactors the provisioning tests, so that the predicate-matching tests
are split out from the event handling & deprovisioning tests.

See: #6660

Source-commit: gravitational/teleport.e@5900a37b17731460d5837ee11a602ab3a1e0505c
Source-pull-requests: gravitational/teleport.e#6742