Remote cluster role filtering PoC
Prior to this change, when reviewing an Access Request for a remote resource, Teleport would just
give up and return all of the user's `search_as` roles, leaving the decision of which roles to
select entirely up to the requester.

This change adds support for calling `GetAccessCapabilities()` from a remote cluster, and having
the local cluster work out what local roles the request requires, and then mapping those roles
back to a subset of remote roles supplied by the caller.