[tctl] Add support for debug sock endpoint with fallback in top
With this change `tctl top` will now attempt to connect to the local
socket debug endpoint before defaulting to the default diag addr.

If the user specifies the endpoint explicitly then the command will
always attempt the given address regardless of success. This matches
previous behavior.

The implementation is not concurrent as only two options exist, both
of which are local and not expected to have long round trips.

Closes: #56673

changelog: `tctl top` now supports the local unix sock debug endpoint.