MWI: Add documentation for bound keypair joining (#56604)
* MWI: Add documentation for bound keypair joining

This adds documentation for bound keypair joining, including a
user-facing deployment guide, an admin guide (with architecture
information), and the usual provision token details.

* TODOs for deployment guide, add deployment guide to method list

* Finish deployment guide and make small tweaks to other pages

* Update tbot CLI and configuration reference

* Fix docs lints

* Add notes about new lock targets for bots

* Bound keypair guide feedback

Various bits of review feedback, and attempt to wordsmith on the
introduction paragraphs.

* Fix review feedback and missing references

Adds some missing references and addresses review feedback.

* Address review feedback, add missing detail

Adds missing detail, especially around decoupled keypair from certs.

* Further clarify what automatic recovery is

* Consistently refer to "bot hosts" rather than nodes

* Address more review feedback

- Retitle the reference guide
- Link to preregistered keys in the guide

* Add guides section to reference page

* Remove admin guide from this PR

This removes the admin guide from this PR and replaces references to
it with placeholders with commented TODOs. The admin guide will be
reintroduced in a separate PR stacked onto this one.

* Fix docs lint

* Fix Machine ID -> Machine & Workload Identity naming

* Update more Machine & Workload Identity references