[v16] Introduce modulestest package (#56779)
* Introduce modulestest package (#56721)

The new package is meant to house a test friendly implementation
of modules.Modules. Due to the pervasive and global nature of
modules, this package also contains a copy of of
modules.SetTestModules. In a perfect world this function wouldn't
be needed and modules would be injected where required, however,
that would take considerable effort. The existing test modules
are temporarily left in place until the enterprise tests can be
updated to use modulestest.

* Follow up to modulestest conversions (#56812)

Update to https://github.com/gravitational/teleport/pull/56721.

Fixes hardware key tests that relied on modifying a pointer to the
globally set modules. Tests now set modules multiple times instead.

Updates the golangci-lint rules to allow enterprise tests to
consume the modulestest package.