MWI: Fall back to registering without an existing auth client
This tweaks joining logic to allow clients that provided an existing
auth client (i.e. bots, used to verify bot instance IDs between cert
refreshes) to fall back to not using that auth client if said client
appears to be broken.

It attempts to perform a ping, and if the ping fails, proceeds as if
no existing client was provided.