[tctl] Add support for debug sock endpoint with fallback in top (#56886) (#57026)
* [tctl] Add support for debug sock endpoint with fallback in top

With this change `tctl top` will now attempt to connect to the local
socket debug endpoint before defaulting to the default diag addr.

If the user specifies the endpoint explicitly then the command will
always attempt the given address regardless of success. This matches
previous behavior.

The implementation is not concurrent as only two options exist, both
of which are local and not expected to have long round trips.

Closes: #56673

changelog: `tctl top` now supports the local unix sock debug endpoint.

* refactor: Simplify top address behaviour based on PR feedback

* Refactor top to reuse debug client for metrics.

* fix linting issues

* Update docs to match support for unix sockets

* Fix imports

* refactor debug service

* clean up error handling in top command

* Add footer display for addr

* Docs for diag/client

* update docs and add addr parser tests

* More docs fixes

* Fix linting after rebase

* Better godocs, more URL test cases, fixed user facing errors

* Fix err trace wrapping semantics

---------

Co-authored-by: Tim Ross <tim.ross@goteleport.com>