MWI: Fall back to registering without an existing auth client (#56927)
* MWI: Fall back to registering without an existing auth client

This tweaks joining logic to allow clients that provided an existing
auth client (i.e. bots, used to verify bot instance IDs between cert
refreshes) to fall back to not using that auth client if said client
appears to be broken.

It attempts to perform a ping, and if the ping fails, proceeds as if
no existing client was provided.

* Explicitly check for an expired identity during renewal

This adds an explicit check for an expired identity during renewal.
If expired, the existing auth client will not be used.

* Revert changes to the join client

* Linter appeasement

* Tweak warning message on expired identity renewal

* Add note about future improvements and tweak logged warning