Add OTP fallback for SSHAgentMFAWebSessionLogin (#57133)
* Add OTP fallback for SSHAgentMFAWebSessionLogin

In the case Teleport `second_factor` is set to `on`,
it is possible for a user to only have OTP configured.
Prior to this commit this would result in a auth fail
as the created challenge only supported TOTP.

Server side would recieve an incomplete `AuthenticateWebUserRequest`
object, which fails validation via `CheckAndSetDefaults` and
as such only a failed login audit event was emitted with no
additional logs.

This mode of failure can be reproduced with
`tsh bench -d web sessions --auth=local`.

changelog: Fixed fallback for web login when second factor is set to `on` but only OTP is configured.

* improve errors for unsupported web login mfa challenge

* Update lib/client/weblogin.go

Co-authored-by: Edoardo Spadolini <edoardo.spadolini@goteleport.com>

* Update lib/client/weblogin.go

Co-authored-by: Edoardo Spadolini <edoardo.spadolini@goteleport.com>

* unexport SSHAgentMFAWebSessionLogin and SSHAgentLoginWeb

* update sshAgentLoginWebCreateSession signature

---------

Co-authored-by: Edoardo Spadolini <edoardo.spadolini@goteleport.com>