Remove testing package from lib/services dependency tree
Most of the places testing crept into production code was via the
lib/services/suite package. The package was removed, and its
helpers were moved into various test pacakges that consumed them.
While this does result in a bit of duplication, the little bit of
copying is worth removing headache that comes with test only
helper packages.

The other places that lib/services brings in the testing package
is via a transitive dependency of sigstore-go:

```
 goda tree "reach(github.com/sigstore/sigstore-go/pkg/tlog...:all, testing)"
  ├ github.com/sigstore/sigstore-go/pkg/tlog
    ├ github.com/sigstore/rekor/pkg/types/dsse/v0.0.1
      └ github.com/in-toto/in-toto-golang/in_toto
        └ github.com/secure-systems-lab/go-securesystemslib/signerverifier
    ├ github.com/sigstore/rekor/pkg/types/intoto/v0.0.2
      └ github.com/in-toto/in-toto-golang/in_toto ~
    └ github.com/sigstore/sigstore-go/pkg/root
      └ github.com/sigstore/timestamp-authority/pkg/verification
        ├ github.com/digitorus/pkcs7
        └ github.com/digitorus/timestamp
          └ github.com/digitorus/pkcs7 ~
```
There is some ongoing work upstream to resolve this, and it will
be addressed by updating `github.com/digitorus/pkcs7` when remediated.

Updates #51023.