[AWSIC] Fix flakey integration test (#6964)
* [AWSIC] Fix flakey integration test

The Identity Center plugin uses a leader lock to ensure that only one integration
service is sunning at a time. If this lock is still held when the Identity Center
service tries to start, the new servce will only try to acquire the lock once a
minute. This is far too long for most of our tests, which will time out waiting to
acquire the lock.

This change adds a configurable retry interval that can be used to override the
default settings during tests.

* Address review feedback

* Finish comment

* Update manager.go

* lint

Source-commit: gravitational/teleport.e@b18422cec529ffaed03730edef9326bb8ee0c65d
Source-pull-requests: gravitational/teleport.e#6964