buildbox-ng: Add node et al to build web UI and Teleport Connect
Add node and other required tools to buildbox-ng so it can be used to
build the web UI and Teleport Connect. This allows buildbox-ng to
replace buildbox-node - one fewer buildboxes in the end.

Note that Teleport connect must be built via the top-level Makefile
target `release-connect` as the Makefile sets up the environment for
cross-compiling properly, which is needed by `node-gyp` for building
native code for the target.

`python3` is installed for `node-gyp`.

`binaryen` is installed for `wasm-opt` as `wasm-pack` 0.12.1 cannot
install it on ARM64, so we just install it globally.

`fpm` (via `ruby` `gem`) is installed as `electron-builder` tries to
install a amd64 version on an arm64 host, which fails for obvious
reasons. We set the `USE_SYSTEM_FPM` env var so `electron-builder` uses
this version.

Set `HOME` as node and associated tooling wants to create stuff in there
for caching. We point it to `/tmp/build` as that is already set up as
the base cache directory for the buildbox for when you want to be able
to persist caches across invocations.