[tctl/top] Parse teleport config file
Add config file parsing to `tctl top` command,
this normally is done on client configuration
but since the `top` client does not make use
of the common tctl client it previously did not
support the local config file.
This this change it is now possible to use a custom
datadir for the debug socket.

changelog: tctl top respects local teleport config file