ui: differentiate desktop certs on the audit log page
Desktop access uses Teleport's user CA to issue certificates, so
the cert.create event in the audit log renders the same for desktop
certificates as for user logins.

In the future, we should probably use a different audit event entirely
for when desktop certs are issued, since there's some Windows-specific
metadata in the cert that should probably be captured in the audit log.

For now, we change the title of the event in the audit log table to
it's easier to spot the difference between login certs and desktop
certs.