[v18] Add "How it works" sections to 16 docs guides (#57228)
* Add "How it works" sections to 16 docs guides

In the Teleport documentation, we are standardizing our how-to guides on
including a "How it works" section after the introduction and before the
prerequisites. This ensures that each guide includes an architectural
overview, which allows the reader to form a mental model of the
instructions before following them.

This change adds "How it works" sections to 16 docs guides related to
access controls, infrastructure as code, and Helm deployments.

* Respond to hugoShaka feedback

- Correct the "How it works" section in the Headless WebAuthn guide.
- Expand the "How it Works" section in the EKS, AKS, and GKE deployment
  guides.
- Add a clarification to the "How it works" section of the agentless SSH
  guide.