Deprecate MakeTestServer in favor of NewTeleportProcess
NewTeleportProcess was introduced to provide the same functionality
without the dependency on testing.T. It is the callers responsibility
to handle errors and clean up the returned process accrodingly. This
helps to ensure that the test package provides only an easy way to
configure and start a teleport process, but makes no assumptions
about what environment the process is being launched in.

The deprecated MakeTestServer still exists because it is referenced
in teleport.e. Once updated to use NewTeleportProcess the legacy
function will be removed.