Make LockWatcher forward deleted lock to subscribers
This change makes the LockWatcher forward its own cached copy of a lock to itd subscribers on
a delete event, rather than the truncated Resounce Header that usually accompanies a delete.

The motivating example for this change is the AWS Identity Center integration. When a lock is
deleted, the IC integration need to know the lock target(s) in order to trigger a permission
assignment recalculation and possible update, and it can't get this information from the vestigial
Resource Header.

The LockWatcher already contains a lock-ID-to-lock-resource mapping internally, so having it supply
its copy of the deleted lock resource to its subscribers seemed the most straightforward way to get
that lock info to the IC integration.