Edit the Teleport Cloud architecture guide (#57455)
Closes #57135

Mention the multi-region HA SLA. Link to the pricing guide PDF, which
contains more information about the SLA, so we can keep all information
in a single place. While there's an exiting variable for the standard
Cloud SLA, this is the only place we mention the mult-region HA SLA, so
adding a variable is premature.