librdpclient: add a timestamp to auto-generated header file (#57357)
A gotcha with our Rust/CGo setup is that if you make a Rust-only
change during development and rebuild Teleport, Go will not realize
that the static library changed and will not re-link it.

Go will, however, detect a change to the header file. Add a timestamp
to the header so that it changes on rebuild, which will be enough to
force Go to link the updated library.

Also rename the file while we're here. Teleport no longer uses
rdp-rs, so librdprs.h is not a good name for the file.