Add Vale to the Lint (Docs) workflow (#57175)
* Add Vale to the Lint (Docs) workflow

Install and run Vale in the Lint (Docs) workflow in order to ensure that
the documentation has a consistent style. This approach downloads a
release from the Vale GitHub repository, extracts a precompiled binary,
and runs it on all docs files. It assumes that all Vale warnings and
errors are already resolved, so that any new ones would have been
introduced by the target pull request.

This is an alternative to using reviewdog, which was behind a security
incident (see #53162 for context). While reviewdog can inspect the diff
introduced by a PR and limit its checks to changed lines, this change
does not require reliance on an insecure dependency.

* Respond to camscale feedback

- Use the `--repo` flag instead of the `checkout` GHA job to download
  a Vale release.
- Perform an integrity check on the Vale archive.

* Respond to fheinecke feedback

Make sure environment variables are quoted.

* Print a note before Vale workflow runs

Indicate that it is possible to ignore Vale rules. This way, docs
authors do not need to be blocked on edge cases.