Improve tsh kubectl traces
The tracing support for tsh kubectl was in a semi-working, but mostly
not functional state. The spans were disjoint instead of all being
associated with a singular SpanContext partly because we didn't pass
the SpanContext from parent to child process and partly due to kubectl
limitations. This addresess those limitations by passing the SpanContext
as an environment variable from the parent tsh process to the child. This
gets all Teleport originated spans unified in the same Span, however, the
kubectl originated spans - by way of us setting a custom http.RoundTripper
were not grouped because kubectl uses a context.TODO for all requests. To
work around this the custom http.RoundTripper was updated to take a
parent context and inject it into the http.Request before it is passed
along to the tracing machinery.