Reduce the number of custom set implementations (#57480)
* Reduce the number of custom set implementations

* Move set to a dedicated package

lib/utils is a large package and we don't want to force all packages
that need a set to import everything else in there.

A small alias remains in lib/utils to avoid breakign teleport.e.

* Fix tests

Some tests were depending on sorted elements, others cared about
the difference between a nil slice and a non-nil but empty slice.