[v18] ci: Move lint-ui workflow into common lint workflow (#57477)
* [v18] ci: Move lint-ui workflow into common lint workflow

Move the `lint-ui.yaml` workflow into the common `lint.yaml` workflow so
we can use the existing setup with the `dorny/paths-filter` action. This
allows us to remove the bypass workflow and fixes the `merge_group`
trigger to also use paths. The original workflow had paths specified for
the `merge_group` event, but that even does not take a `paths` field.

This fixes a bug where the lint-ui workflow does not run on a PR, but
does run in the merge queue, even if none of the UI workflow files
changed.

* [v18] Fix buildbox version for lint-rfd

Change the buildbox container tag in the lint-rfd job from `:teleport17`
to `:teleport18`, as it should be that on `branch/v18`.