Move pages out of the `admin-guides/api` section (#57440)
We are ensuring that all docs pages about Teleport features fall under a
product-focused section. Move the API how-to guides to Zero Trust Access
except for the Access Request pluging guide, which requires Identity
Governance.

Move the Migrate Plans page to the root of the docs since it is a
general-purpose document for all users.