Fix CRLs in the backend
In #55429 we added an empty certificate revocation list to the
TLS keypairs in the certificate_authority backend resource.

The intent was that the CRL would be signed by the keypair it is
assocaited with, but the code mistakenly used a function that would
find _any_ suitable keypair, which could create CRLS signed by the
wrong key.

This commit corrects both the initial CRL generation as well as
any existing CRLs that have invalid signatures.