Add some extra warnings re: CRL generation
In clusters with multiple active signers (like those using HSMs or
KMS for private keys), the only safe way to get a CRL is to use
the one that's already present in the certificate_authority resource.

Generating a CRL with GenerateCertAuthorityCRL will give you a valid
CRL, but you have no control over which certificate signs it, which
is likely to cause revocation failures later on.