AWS Roles Anywhere: endpoint to validate name
When the user is configuring the integration, one of the steps is to
give a name to the integration.
After giving the name, the user is asked to run a script.
Only after, the integration is created on Teleport side.

This can cause a bad UX if the user picks an integration name that is
not valid.

This endpoint is going to be called from the UI when the user enters the
integration name.

It ensures the integration does not exist yet, and that it is a valid
name.