winpki: prevent CRL common names from overflowing (#57907)
The addition of the issuer SKID to the CN for our CRLS  has
caused long clusters to exceed the limit of how long a CN
can be.

For database access, we rely on the user to run certutil
commands to publish the CRL. While we expect the limit to
be 64 characters, we observe that certutil starts truncating
the CN as soon as it exceeds 51 characters (which causes it
to get imported in a different location from what the certificate
references).