Move remaining Admin Guides content to ZTA (#57943)
Backports #57439 and #57440

We are ensuring that all docs pages about Teleport features fall under a
product-focused section. Move the API how-to guides to Zero Trust Access
except for the Access Request pluging guide, which requires Identity
Governance.

Move the Migrate Plans page to the root of the docs since it is a
general-purpose document for all users.

The guides in the `admin-guides/management` section describe ZTA
features, and we are organizing the docs so that all pages fall under a
product-focused category. Move them to ZTA.