Restructure Teleport RBAC guidance (#58074)
Backports #57921

Closes #21313
Closes #15016
Closes #39241

Beginner-level Teleport RBAC guidance is currently in Access Controls,
where it is difficult to find within more advanced guides on RBAC
topics. This change restructures the RBAC documentation so
beginner-level information is more visible:

**Change the Access Controls section to a section on authentication:**
The guides in this section are mostly about configuring authentication,
not about access controls in general. Meanwhile, there are guides
throughout the docs on access controls in general, leaving the Access
Controls section with an ambiguous purpose.

**Remove the Dual Authorization guide:** Use the Access Requests landing
page to describe the connection to Dual Authorization. This lets us
avoid repeating Access Request plugin content in the Access Controls
section and makes it easier to organize the Access Controls guides.

**Add a beginner-level section on RBAC:** Add the guides to labeling and
role templates here, along with a getting started guide with a
step-by-step demo example of setting up RBAC, aimed at beginners.