AWS IAM Roles Anywhere: ignore empty Trust Anchor on update
When updating the integration (e.g., when changing the profile filters),
we were expecting to receive the Trust Anchor ARN, in order to perform a
full upgrade on the resource.

However, this is not desirable in the UI.
Given that, the endpoint will now only update the profile sync fields.