Add ability to interact with cluster state backend directly
The teleport backend command has been extended with four new
subcommands: get, ls, rm, and edit. These commands can be used
to retrieve, remove, or modify backend items without having to
use additional tools. This is particularly useful for the
Firestore backend since it encodes all keys and values as base64
which makes using the tools Firestore provides to inspect the
cluster state difficult.

These new backend commands use the same Teleport configuration
file that the Auth server does and connect directly to the
backend the same way the Auth server does. This means that in
order for these tools to be useful they need to be executed from
a location and by a user that has equivalent permissions that the
Auth server does. Due to which these commands are hidden and
should only be needed to aid in backend investigations as a last
ditch measure.


```bash

$ teleport backend -h
usage: teleport backend <command> [<args> ...]

Commands for managing backend data.

Flags:

Commands:
  backend get   Retrieves a single item from the cluster state backend.
  backend rm    Removes a single item from the cluster state backend.
  backend ls    Lists the keys in the cluster state backend.
  backend edit  Modify a single item from the cluster state backend.
  backend clone Clones data from a source to a destination backend.

```