AWS IAM Roles Anywhere: fix stats endpoint
The integration/<name>/stats endpoint can be used for both AWS OIDC and
AWS IAM Roles Anywhere integrations.

It is used by the UI to show the summary of the integration.

When used for the Roles Anywhere, it was also asking for the number of
deployed database services.
This feature is not yet implemented in the Roles Anywhere integration,
and so it's not possible to count the deployed services.

This PR changes the flow to only count the services if this is an AWS
OIDC integration.