AWS IAM Roles Anywhere: Test connection ignores the Profile sync (#58256)
* AWS Roles Anywhere: Test connection ignores the Profile sync

When the user is setting up the AWS IAM Roles Anywhere integration,
they will test the connection.
This test connection (aka ping) will let the user know if the
credentials are working and how many profiles they have that they can
use.

Before this PR, we were also counting the Profile that is used to sync
profiles.
Even though, this is a valid Profile, it will not be added as an AWS
App, it is only used to sync profiles.

After this PR, we are no longer including it.
The UI, when there's no profiles, will also let the user know about this
and ask them to create profiles.

* review pt1