Add Owners and Roles to new filter message
The frontend requires adding owners/roles to the filter (along side
search) similar to how we send kinds in unified resources. we still
check against roles/owners in the search bar as well because someone
should be able to select the owner filter in the page OR just type it
in. the search bar is like the "even more detailed" filter, as we will
check the filters owners/roles first, then filter it down further is
search exists