[v18] Add links between Azure protection guides (#57650)
* Add links between Azure protection guides

Closes #54091

There are two guides for protecting Azure with Teleport, and each uses a
different set of Teleport features. Add links between the two guides.

* Respond to flyinghermit feedback

Provide more nuance in describing the two available integration
approaches. This includes adding a "How it works" section to the Azure
Portal guide, fulfilling one expectation we are standardizing for how-to
guides in general.

* updte how it works

* Apply minor grammar and style changes

---------

Co-authored-by: flyinghermit <sshah@goteleport.com>