[loadtest] Add batch create resource Helm chart (#58579)
This Helm chart is for internal testing purposes only and
is subject to change.

It allows spawning of arbitary number of batch k8s jobs to
run commands on the Teleport cluster such as creating resources.

For now only app kind resource is used but this could be extended
in the future. It makes use of `tbot` to authenticate the batch jobs.

The resources are not deleted each time as currently the intent is
to scale the number of resources in the backend while testing and
tear down the cluster once done.