Enable Docusaurus tagging in docs frontmatter
Replace the `labels` frontmatter field with one called `tags`.
Docusaurus supports tag-based navigation out of the box with the `tags`
field. The original plan was to implement a separate system that allows
for filtering by multiple tags - which Docusaurus does not - but it
makes sense to enable the default approach as a first iteration.